Typing updates
