+++ app/resources/types/warp.ts
@@ -16,13 +16,4 @@ type WarpLengthData={
 }
 
 
-type grid=color[][]
-type gridName='weave' | 'tieup'|'warp'|'treadle'|'weft'
-
-type GridSet={
-    warpGrid:grid, 
-    tieUpGrid:grid, 
-    treadleGrid:grid,
-    weave?:grid
-}
